+++ src/alg/cell/cell_refinement.c
@@ -27,7 +27,7 @@ void refine_cell( struct cell_node *cell, uint32_t *free_sv_positions, uint32_t 
 
     int number_of_hilbert_shape;
 
-    double cell_center_x    = cell->center_x,
+    float cell_center_x    = cell->center_x,
             cell_center_y   = cell->center_y,
             cell_center_z   = cell->center_z,
             cell_half_side    = cell->half_face_length,
@@ -1164,8 +1164,8 @@ void simplify_refinement( struct transition_node *transition_node ) {
 }
 
 void set_refined_cell_data(struct cell_node* the_cell, struct cell_node* other_cell,
-                           double face_length, double half_face_length,
-                           double center_x, double center_y, double center_z,
+                           float face_length, float half_face_length,
+                           float center_x, float center_y, float center_z,
                            uint64_t  bunch_number, uint32_t *free_sv_positions,
                            uint32_t **refined_this_step) {
